refactor: cleaning up the test-utils pkg
