+++ packages/api/test/__fixtures__/sdk/response-title-quirks/index.ts
@@ -2,7 +2,7 @@ import type * as types from './types';
 import type { ConfigOptions, FetchResponse, HTTPMethodRange } from '@api/core';
 import Oas from 'oas';
 import APICore from '@api/core';
-import definition from '@api/test-utils/src/fixtures/definitions/response-title-quirks.json';
+import definition from '@api/test-utils/definitions/response-title-quirks.json';
 
 class SDK {
   spec: Oas;
