+++ packages/api/vitest-smoketest.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     include: ['**/smoketest.test.ts'],
-    setupFiles: ['./test/helpers/vitest.matchers.ts'],
+    setupFiles: ['@api/test-utils/vitest.matchers.ts'],
     testTimeout: 20000,
   },
 });
