+++ packages/core/test/lib/prepareAuth.test.ts
@@ -1,6 +1,6 @@
 import type { OASDocument } from 'oas/dist/rmoas.types';
 
-import loadSpec from '@api/test-utils/src/load-spec';
+import loadSpec from '@api/test-utils/load-spec';
 import Oas from 'oas';
 import { describe, beforeAll, it, expect } from 'vitest';
 
@@ -167,7 +167,7 @@ describe('#prepareAuth()', () => {
     let securityMultipleOas;
 
     beforeAll(async () => {
-      authQuirksOas = await loadSpec(require.resolve('@api/test-utils/src/fixtures/definitions/auth-quirks.json'));
+      authQuirksOas = await loadSpec(require.resolve('@api/test-utils/definitions/auth-quirks.json'));
       securityMultipleOas = await loadSpec('@readme/oas-examples/3.0/json/security-multiple.json');
     });
