+++ packages/core/vitest.config.ts
@@ -3,6 +3,6 @@ import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
   test: {
-    setupFiles: ['@api/test-utils/src/vitest.matchers.ts'],
+    setupFiles: ['@api/test-utils/vitest.matchers.ts'],
   },
 });
