+++ packages/test-utils/definitions/auth-quirks.json
