+++ packages/test-utils/fixtures/owlbert-shrub.png
