+++ packages/test-utils/tsconfig.json
@@ -7,5 +7,5 @@
     "noEmit": true,
     "noImplicitAny": true
   },
-  "include": ["./src/**/*"]
+  "include": ["./**/*"]
 }
